QuestionController, QuestionService, QuizController et QuizService

+++ spring_pg_starter/src/main/java/org/amu/repository/QuestionRepository.java
@@ -2,5 +2,5 @@ package org.amu.repository;
 
 import org.amu.model.Question;
 
-public interface QuestionRepository extends CrudRepository<Question, Integer>{
+public interface QuestionRepository extends CrudRepository<Question, Long>{
 }
